Simplified `TransformerIterator`
Dhii iterators update their `valid` status after they `_loop()`. Looping, for this implementation, means transforming the iteration.
Since PHP calls `next()` before `valid()` when iterating using `foreach`, transformations will be applied before checking if the inner iterator is valid or not.
This results in the transformer getting called with a `null` source.

Since this implementation is meant to transform the values of an iterator using a single transformer, the keys are now non-transformable.
If consumers wish to also transform keys, the new `MapTransformer` is much more flexible.